\name{LocationWQParameters}
\alias{LocationWQParameters}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
LocationWQParameters
}
\description{
Returns a list all available WQ parameters for a given location

This list can then be passed directly into AQMultiExtract if all parameters are required
}
\usage{
LocationWQParameters("AQlocationid")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  AQlocationID
}
\author{
Darren Gerretzen
}
