\name{sample_flowsite_mapping}
\alias{sample_flowsite_mapping}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
sample_flowsite_mapping
}
\description{
Returns mapping between sample and flow sites.

The mapping is a copy of the '.NERMN_Lab_Flow_Configuration' stored in Objective - A2741215
}
\usage{
sample_flowsite_mapping()
}
